optimisation of home page (extraction of navbar and footer)
